Feature 01

+++ src/main/java/com/carryit/base/besttmwuu/entity/imsEweiShopSnsPost.java
@@ -21,7 +21,7 @@ public class imsEweiShopSnsPost {
 
     private String voice;
 
-    private Integer createtime;
+    private Long createtime;
 
     private Integer replytime;
 
@@ -131,11 +131,11 @@ public class imsEweiShopSnsPost {
         this.voice = voice == null ? null : voice.trim();
     }
 
-    public Integer getCreatetime() {
+    public Long getCreatetime() {
         return createtime;
     }
 
-    public void setCreatetime(Integer createtime) {
+    public void setCreatetime(Long createtime) {
         this.createtime = createtime;
     }
